+++ app/src/main/java/com/example/dictionaryappdemo/activity/DetailActivity.kt
@@ -1,7 +1,8 @@
-package com.example.dictionaryappdemo
+package com.example.dictionaryappdemo.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.dictionaryappdemo.R
 
 class DetailActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
