+++ app/src/main/java/com/example/dictionaryappdemo/activity/MainActivity.kt
@@ -1,9 +1,11 @@
-package com.example.dictionaryappdemo
+package com.example.dictionaryappdemo.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.dictionaryappdemo.R
+import com.example.dictionaryappdemo.adapter.LetterAdapter
 import com.example.dictionaryappdemo.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
