+++ app/src/main/java/com/example/dictionaryappdemo/adapter/LetterAdapter.kt
@@ -1,6 +1,5 @@
-package com.example.dictionaryappdemo
+package com.example.dictionaryappdemo.adapter
 
-import android.content.Context
 import android.os.Build
 import android.view.LayoutInflater
 import android.view.View
@@ -9,65 +8,48 @@ import android.view.accessibility.AccessibilityNodeInfo
 import android.widget.Button
 import androidx.annotation.RequiresApi
 import androidx.recyclerview.widget.RecyclerView
+import com.example.dictionaryappdemo.R
 
 /**
- * Adapter for the [RecyclerView] in [DetailActivity].
+ * Adapter for the [RecyclerView] in [MainActivity].
  */
-class WordAdapter(private val letterId: String, context: Context) :
-    RecyclerView.Adapter<WordAdapter.WordViewHolder>() {
+class LetterAdapter :
+    RecyclerView.Adapter<LetterAdapter.LetterViewHolder>() {
 
-    private val filteredWords: List<String>
+    // Generates a [CharRange] from 'A' to 'Z' and converts it to a list
+    private val list = ('A').rangeTo('Z').toList()
 
-    init {
-        // Retrieve the list of words from res/values/arrays.xml
-        val words = context.resources.getStringArray(R.array.words).toList()
-
-        filteredWords = words
-            // Returns items in a collection if the conditional clause is true,
-            // in this case if an item starts with the given letter,
-            // ignoring UPPERCASE or lowercase.
-            .filter { it.startsWith(letterId, ignoreCase = true) }
-            // Returns a collection that it has shuffled in place
-            .shuffled()
-            // Returns the first n items as a [List]
-            .take(5)
-            // Returns a sorted version of that [List]
-            .sorted()
-    }
-
-    class WordViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
+    /**
+     * Provides a reference for the views needed to display items in your list.
+     */
+    class LetterViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
         val button = view.findViewById<Button>(R.id.button_item)
     }
 
-    override fun getItemCount(): Int = filteredWords.size
+    override fun getItemCount(): Int {
+        return list.size
+    }
 
     /**
      * Creates new views with R.layout.item_view as its template
      */
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): WordViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): LetterViewHolder {
         val layout = LayoutInflater
             .from(parent.context)
             .inflate(R.layout.item_view, parent, false)
-
         // Setup custom accessibility delegate to set the text read
         layout.accessibilityDelegate = Accessibility
-
-        return WordViewHolder(layout)
+        return LetterViewHolder(layout)
     }
 
     /**
      * Replaces the content of an existing view with new data
      */
-    override fun onBindViewHolder(holder: WordViewHolder, position: Int) {
-
-        val item = filteredWords[position]
-        // Needed to call startActivity
-        val context = holder.view.context
-
-        // Set the text of the WordViewHolder
-        holder.button.text = item
-
+    override fun onBindViewHolder(holder: LetterViewHolder, position: Int) {
+        val item = list.get(position)
+        holder.button.text = item.toString()
     }
+
     // Setup custom accessibility delegate to set the text read with
     // an accessibility service
     companion object Accessibility : View.AccessibilityDelegate() {
@@ -81,7 +63,7 @@ class WordAdapter(private val letterId: String, context: Context) :
             // accessibility service announces "double tap to activate".
             // If a custom string is provided,
             // it announces "double tap to <custom string>".
-            val customString = host?.context?.getString(R.string.look_up_word)
+            val customString = host?.context?.getString(R.string.look_up_words)
             val customClick =
                 AccessibilityNodeInfo.AccessibilityAction(
                     AccessibilityNodeInfo.ACTION_CLICK,
